requirements.txt
python-telegram-bot==13.15
APScheduler
python-dotenv
urllib3==1.26.16